+++ chevah/compat/capabilities.py
@@ -7,6 +7,7 @@ from __future__ import print_function
 from __future__ import division
 from __future__ import absolute_import
 import os
+import platform
 import sys
 
 
@@ -35,6 +36,27 @@ def _get_os_name():
         raise AssertionError('OS "%s" not supported.' % family)
 
 
+def _get_cpu_type():
+    """
+    Return the CPU type as used in the paver.sh script.
+    """
+    base = platform.processor()
+
+    if not base:
+        base = platform.machine()
+
+    if base == 'aarch64':  # noqa:cover
+        return 'arm64'
+
+    if base == 'x86_64':
+        return 'x64'
+
+    if base == 'i686':
+        return 'x86'
+
+    return base
+
+
 class BaseProcessCapabilities(object):
     """
     Code shared by all `IProcessCapabilities`
@@ -42,3 +64,4 @@ class BaseProcessCapabilities(object):
 
     os_family = os.name
     os_name = _get_os_name()
+    cpu_type = _get_cpu_type()