+++ chevah/compat/testing/constants.py
@@ -6,4 +6,4 @@ from __future__ import print_function
 from __future__ import division
 from __future__ import absolute_import
 
-TEST_NAME_MARKER = u'-cpț'
+TEST_NAME_MARKER = u'-Cpț'