+++ release-notes.rst
@@ -2,7 +2,39 @@ Release notes for chevah.compat
 ===============================
 
 
-0.52.4 - 26/03/2019
+0.55.0 - 12/05/2019
+-------------------
+
+* Updated testing text generator to include upper and lower characters.
+
+
+0.54.1 - 08/05/2019
+-------------------
+
+* Remove TODOs for Solaris/AIX/HPUX as there is no plan to fix them.
+* Fix test case teardown.
+
+
+0.54.0 - 15/04/2019
+-------------------
+
+* Fix reactor debug mode.
+* Fix assertEqual str vs unicode check.
+
+
+0.53.0 - 03/04/2019
+-------------------
+
+* Fix command line argument parsing when using multiprocessiong.
+
+
+0.52.5 - 26/03/2019
+-------------------
+
+* Fix previous base version on Chevah PyPi.
+
+
+0.52.4 - 24/03/2019
 -------------------
 
 * getHomeFolder now always returns a path without the trailing separater.